requirements.txt
Django==1.9.7
django-debug-toolbar==1.4
MySQL-python==1.2.5
sqlparse==0.2.1
grpcio==1.0.1
grpcio-tools==1.0.1
